+++ AnimeVideoDownloader/DesktopDownloader/App.xaml.cs
@@ -1,7 +1,15 @@
 ﻿
+using System.Windows;
+using Serilog;
+
 namespace DesktopDownloader {
 	/// <summary>
 	/// Interaction logic for App.xaml
 	/// </summary>
-	public partial class App { }
+	public partial class App {
+		protected override void OnStartup(StartupEventArgs e) {
+			base.OnStartup(e);
+
+		}
+	}
 }