+++ AnimeVideoDownloader/DesktopDownloader/Data/Emoji.cs
@@ -1,0 +1,6 @@
+﻿namespace DesktopDownloader.Data {
+	public class Emoji {
+		public const string Done = "✔️";
+		public const string InProgress = "⌛";
+	}
+}